requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
sqlalchemy
pydantic
alembic